enhance: Use GCPolicy to refcount instead of GCManager
